requirements.txt
# Microsoft Agent Framework - Python Dependencies

# Core Agent Framework
agent-framework>=0.1.0

# Azure OpenAI and Authentication
azure-identity>=1.15.0
openai>=1.12.0

# Additional utilities
python-dotenv>=1.0.0
